+++ source/extensions/filters/http/jwt_authn/verifier.cc
@@ -71,7 +71,7 @@ private:
 };
 
 // base verifier for provider_name, provider_and_audiences, and allow_missing_or_failed.
-class BaseVerifierImpl : public Verifier {
+class BaseVerifierImpl : public Logger::Loggable<Logger::Id::jwt>, public Verifier {
 public:
   BaseVerifierImpl(const BaseVerifierImpl* parent) : parent_(parent) {}
 
@@ -112,7 +112,7 @@ public:
 
   void verify(ContextSharedPtr context) const override {
     auto& ctximpl = static_cast<ContextImpl&>(*context);
-    auto auth = auth_factory_.create(getAudienceChecker(), provider_name_, false);
+    auto auth = auth_factory_.create(getAudienceChecker(), provider_name_, false, false);
     extractor_->sanitizePayloadHeaders(ctximpl.headers());
     auth->verify(
         ctximpl.headers(), ctximpl.parentSpan(), extractor_->extract(ctximpl.headers()),
@@ -156,16 +156,17 @@ private:
 // Allow missing or failed verifier
 class AllowFailedVerifierImpl : public BaseVerifierImpl {
 public:
-  AllowFailedVerifierImpl(const AuthFactory& factory, const Extractor& extractor,
+  AllowFailedVerifierImpl(const AuthFactory& factory, const JwtProviderList& providers,
                           const BaseVerifierImpl* parent)
-      : BaseVerifierImpl(parent), auth_factory_(factory), extractor_(extractor) {}
+      : BaseVerifierImpl(parent), auth_factory_(factory), extractor_(Extractor::create(providers)) {
+  }
 
   void verify(ContextSharedPtr context) const override {
     auto& ctximpl = static_cast<ContextImpl&>(*context);
-    auto auth = auth_factory_.create(nullptr, absl::nullopt, true);
-    extractor_.sanitizePayloadHeaders(ctximpl.headers());
+    auto auth = auth_factory_.create(nullptr, absl::nullopt, true, true);
+    extractor_->sanitizePayloadHeaders(ctximpl.headers());
     auth->verify(
-        ctximpl.headers(), ctximpl.parentSpan(), extractor_.extract(ctximpl.headers()),
+        ctximpl.headers(), ctximpl.parentSpan(), extractor_->extract(ctximpl.headers()),
         [&ctximpl](const std::string& name, const ProtobufWkt::Struct& payload) {
           ctximpl.addPayload(name, payload);
         },
@@ -181,12 +182,48 @@ public:
 
 private:
   const AuthFactory& auth_factory_;
-  const Extractor& extractor_;
+  // const Extractor& extractor_;
+  const ExtractorConstPtr extractor_;
+};
+
+class AllowMissingVerifierImpl : public BaseVerifierImpl {
+public:
+  AllowMissingVerifierImpl(const AuthFactory& factory, const JwtProviderList& providers,
+                           const BaseVerifierImpl* parent)
+      : BaseVerifierImpl(parent), auth_factory_(factory), extractor_(Extractor::create(providers)) {
+  }
+
+  void verify(ContextSharedPtr context) const override {
+    ENVOY_LOG(debug, "Called AllowMissingVerifierImpl.verify : {}", __func__);
+
+    auto& ctximpl = static_cast<ContextImpl&>(*context);
+    auto auth = auth_factory_.create(nullptr, absl::nullopt, false /* allow failed */,
+                                     true /* allow missing */);
+    extractor_->sanitizePayloadHeaders(ctximpl.headers());
+    auth->verify(
+        ctximpl.headers(), ctximpl.parentSpan(), extractor_->extract(ctximpl.headers()),
+        [&ctximpl](const std::string& name, const ProtobufWkt::Struct& payload) {
+          ctximpl.addPayload(name, payload);
+        },
+        [this, context](const Status& status) {
+          onComplete(status, static_cast<ContextImpl&>(*context));
+        });
+    if (!ctximpl.getCompletionState(this).is_completed_) {
+      ctximpl.storeAuth(std::move(auth));
+    } else {
+      auth->onDestroy();
+    }
+  }
+
+private:
+  const AuthFactory& auth_factory_;
+  const ExtractorConstPtr extractor_;
 };
 
 VerifierConstPtr innerCreate(const JwtRequirement& requirement,
                              const Protobuf::Map<std::string, JwtProvider>& providers,
-                             const AuthFactory& factory, const Extractor& extractor,
+                             const AuthFactory& factory,
+                             const std::vector<std::string> parent_provider_names,
                              const BaseVerifierImpl* parent);
 
 // Base verifier for requires all or any.
@@ -214,10 +251,41 @@ class AnyVerifierImpl : public BaseGroupVerifierImpl {
 public:
   AnyVerifierImpl(const JwtRequirementOrList& or_list, const AuthFactory& factory,
                   const Protobuf::Map<std::string, JwtProvider>& providers,
-                  const Extractor& extractor_for_allow_fail, const BaseVerifierImpl* parent)
+                  const BaseVerifierImpl* parent)
       : BaseGroupVerifierImpl(parent) {
+    const JwtRequirement* by_pass_type_requirement = nullptr;
+    std::vector<std::string> used_providers;
     for (const auto& it : or_list.requirements()) {
-      verifiers_.emplace_back(innerCreate(it, providers, factory, extractor_for_allow_fail, this));
+      bool is_regular_requirement = true;
+      switch (it.requires_type_case()) {
+      case JwtRequirement::RequiresTypeCase::kProviderName:
+        used_providers.emplace_back(it.provider_name());
+        break;
+      case JwtRequirement::RequiresTypeCase::kProviderAndAudiences:
+        used_providers.emplace_back(it.provider_and_audiences().provider_name());
+        break;
+      case JwtRequirement::RequiresTypeCase::kAllowMissingOrFailed:
+      case JwtRequirement::RequiresTypeCase::kAllowMissing:
+        is_regular_requirement = false;
+        if (by_pass_type_requirement == nullptr ||
+            by_pass_type_requirement->requires_type_case() ==
+                JwtRequirement::RequiresTypeCase::kAllowMissing) {
+          // We need to keep only one by_pass_type_requirement. If both
+          // kAllowMissing and kAllowMissingOrFailed are set, use
+          // kAllowMissingOrFailed.
+          by_pass_type_requirement = &it;
+        }
+      default:
+        break;
+      }
+      if (is_regular_requirement) {
+        verifiers_.emplace_back(
+            innerCreate(it, providers, factory, std::vector<std::string>{}, this));
+      }
+    }
+    if (by_pass_type_requirement) {
+      verifiers_.emplace_back(
+          innerCreate(*by_pass_type_requirement, providers, factory, used_providers, this));
     }
   }
 
@@ -239,10 +307,12 @@ class AllVerifierImpl : public BaseGroupVerifierImpl {
 public:
   AllVerifierImpl(const JwtRequirementAndList& and_list, const AuthFactory& factory,
                   const Protobuf::Map<std::string, JwtProvider>& providers,
-                  const Extractor& extractor_for_allow_fail, const BaseVerifierImpl* parent)
+                  // const Extractor& extractor_for_allow_fail,
+                  const BaseVerifierImpl* parent)
       : BaseGroupVerifierImpl(parent) {
     for (const auto& it : and_list.requirements()) {
-      verifiers_.emplace_back(innerCreate(it, providers, factory, extractor_for_allow_fail, this));
+      verifiers_.emplace_back(
+          innerCreate(it, providers, factory, std::vector<std::string>{}, this));
     }
   }
 
@@ -271,10 +341,19 @@ public:
 
 VerifierConstPtr innerCreate(const JwtRequirement& requirement,
                              const Protobuf::Map<std::string, JwtProvider>& providers,
-                             const AuthFactory& factory, const Extractor& extractor_for_allow_fail,
+                             const AuthFactory& factory,
+                             const std::vector<std::string> parent_provider_names,
                              const BaseVerifierImpl* parent) {
   std::string provider_name;
   std::vector<std::string> audiences;
+  JwtProviderList parent_providers;
+  for (const auto& name : parent_provider_names) {
+    const auto& it = providers.find(name);
+    if (it == providers.end()) {
+      throw EnvoyException(fmt::format("Required provider ['{}'] is not configured.", name));
+    }
+    parent_providers.emplace_back(&it->second);
+  }
   switch (requirement.requires_type_case()) {
   case JwtRequirement::RequiresTypeCase::kProviderName:
     provider_name = requirement.provider_name();
@@ -287,12 +366,14 @@ VerifierConstPtr innerCreate(const JwtRequirement& requirement,
     break;
   case JwtRequirement::RequiresTypeCase::kRequiresAny:
     return std::make_unique<AnyVerifierImpl>(requirement.requires_any(), factory, providers,
-                                             extractor_for_allow_fail, parent);
+                                             parent);
   case JwtRequirement::RequiresTypeCase::kRequiresAll:
     return std::make_unique<AllVerifierImpl>(requirement.requires_all(), factory, providers,
-                                             extractor_for_allow_fail, parent);
+                                             parent);
   case JwtRequirement::RequiresTypeCase::kAllowMissingOrFailed:
-    return std::make_unique<AllowFailedVerifierImpl>(factory, extractor_for_allow_fail, parent);
+    return std::make_unique<AllowFailedVerifierImpl>(factory, parent_providers, parent);
+  case JwtRequirement::RequiresTypeCase::kAllowMissing:
+    return std::make_unique<AllowMissingVerifierImpl>(factory, parent_providers, parent);
   case JwtRequirement::RequiresTypeCase::REQUIRES_TYPE_NOT_SET:
     return std::make_unique<AllowAllVerifierImpl>(parent);
   default:
@@ -319,9 +400,8 @@ ContextSharedPtr Verifier::createContext(Http::HeaderMap& headers, Tracing::Span
 
 VerifierConstPtr Verifier::create(const JwtRequirement& requirement,
                                   const Protobuf::Map<std::string, JwtProvider>& providers,
-                                  const AuthFactory& factory,
-                                  const Extractor& extractor_for_allow_fail) {
-  return innerCreate(requirement, providers, factory, extractor_for_allow_fail, nullptr);
+                                  const AuthFactory& factory) {
+  return innerCreate(requirement, providers, factory, std::vector<std::string>{}, nullptr);
 }
 
 } // namespace JwtAuthn