+++ test/extensions/transport_sockets/tls/ssl_socket_test.cc
@@ -13,7 +13,6 @@
 #include "common/network/transport_socket_options_impl.h"
 #include "common/network/utility.h"
 
-#include "extensions/filters/listener/tls_inspector/tls_inspector.h"
 #include "extensions/transport_sockets/tls/context_config_impl.h"
 #include "extensions/transport_sockets/tls/context_impl.h"
 #include "extensions/transport_sockets/tls/private_key/private_key_manager_impl.h"