+++ tools/proto_sync.py
@@ -120,7 +120,10 @@ def SyncV3Alpha(cmd, src_labels):
       continue
     # Skip unversioned package namespaces. TODO(htuch): fix this to use the type
     # DB and proper upgrade paths.
-    if 'v2' in dst:
+    if 'v1' in dst:
+      dst = re.sub('v1alpha\d?|v1', 'v3alpha', dst)
+      SyncProtoFile(cmd, src, dst)
+    elif 'v2' in dst:
       dst = re.sub('v2alpha\d?|v2', 'v3alpha', dst)
       SyncProtoFile(cmd, src, dst)
     elif 'envoy/type/matcher' in dst:
@@ -149,8 +152,8 @@ def GetImportDeps(proto_path):
         # We can ignore imports provided implicitly by api_proto_package().
         if any(import_path.startswith(p) for p in API_BUILD_SYSTEM_IMPORT_PREFIXES):
           continue
-        # Special case handling for in-built versioning annotations.
-        if import_path == 'udpa/annotations/versioning.proto':
+        # Special case handling for UDPA annotations.
+        if import_path.startswith('udpa/annotations/'):
           imports.append('@com_github_cncf_udpa//udpa/annotations:pkg')
           continue
         # Explicit remapping for external deps, compute paths for envoy/*.